backend/requirements.txt
flask
flask_sqlalchemy
psycopg2
flask_cors
